Add Geocoding to global data panel
